@@ -2,14 +2,18 @@
 name = "shared"
 version = "0.1.0"
 edition = "2021"
-rust-version = "1.74"
+rust-version = "1.80"
 
 [lib]
 crate-type = ["lib", "staticlib", "cdylib"]
 name = "shared"
 
+[[bin]]
+name = "crux_cli"
+required-features = ["cli"]
+
 [features]
-typegen = ["crux_core/typegen"]
+cli = ["crux_core/cli"]
 
 [dependencies]
 crux_core.workspace = true
@@ -19,8 +23,11 @@ uniffi = "0.29.2"
 wasm-bindgen = "0.2.100"
 uuid = { version = "1.7.0", features = ["v4", "serde", "js"] }
 
-[target.uniffi-bindgen.dependencies]
-uniffi = { version = "0.29.2", features = ["cli"] }
+[target.'cfg(not(target_family = "wasm"))'.dependencies]
+uniffi = "0.29.2"
 
-[build-dependencies]
-uniffi = { version = "0.29.2", features = ["build"] }
+[target.'cfg(target_family = "wasm")'.dependencies]
+wasm-bindgen = "0.2.100"
+
+[dev-dependencies]
+insta = { version = "1.43.1", features = ["yaml"] }
